requirements.txt
google-generativeai
groq
edge-tts
moviepy==1.0.3
requests
python-dotenv
imageio-ffmpeg
proglog